pyproject.toml
[project]
name = "ComfyUI-FlashFace"
description = "ComfyUI custom node for FlashFace"
version = "1.3.5"
license = { file = "LICENSE" }
dependencies = [
    "accelerate>=1.10.1",
    "diffusers>=0.33.0",
    "peft>=0.17.0",
    "ftfy>=6.2.0",
    "gguf>=0.17.1",
    "pyloudnorm",
    "gradio>=5.46.0",
    "huggingface-hub>=0.35.0",
    "numpy>=2.3.3",
    "omegaconf==2.3.0",
    "safetensors>=0.6.2",
    "torch==2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.1",
    "easydict>=1.13",
    "pillow>=11.3.0"
]

[project.urls]
Repository = "https://github.com/cold-hand/ComfyUI-FlashFace"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "cold-hand"
DisplayName = "ComfyUI-FlashFace"
Icon = ""
